PersonDetail Servisleri Yazıldı

+++ Assessment.Phonebook.Services/Services/PersonServices.cs
@@ -22,6 +22,7 @@ namespace Assessment.Phonebook.Services.Services
             _mapper = mapper;
             _context = context;
         }
+        #region Person
         public async Task<ResultModel> CreatePerson(PersonDto request)
         {
             var person = new Person
@@ -31,7 +32,7 @@ namespace Assessment.Phonebook.Services.Services
                 Company = request.Company,
                 PersonDetail = new List<PersonDetail>()
             };
-            if (request.PersonDetails!=null)
+            if (request.PersonDetails != null)
             {
                 foreach (var item in request.PersonDetails)
                 {
@@ -41,12 +42,12 @@ namespace Assessment.Phonebook.Services.Services
                         MailAddress = item.MailAddress,
                         PersonId = item.PersonId,
                         Location = item.Location,
-                      
+
                     };
                     person.PersonDetail.Add(personDetail);
                 }
             }
-            
+
             await _context.Persons.AddAsync(person);
             await _context.SaveChangesAsync();
 
@@ -56,14 +57,20 @@ namespace Assessment.Phonebook.Services.Services
         public async Task<ResultModel> DeletePerson(ItemDto request)
         {
             var person = await _context.Persons.Include(person => person.PersonDetail).FirstOrDefaultAsync(x => x.Id == request.Id);
-            foreach (var detail in person.PersonDetail)
+            if (person != null)
             {
-                _context.PersonDetails.Remove(detail);
+                foreach (var detail in person.PersonDetail)
+                {
+                    _context.PersonDetails.Remove(detail);
+                }
+                await _context.SaveChangesAsync();
+                _context.Persons.Remove(person);
+                await _context.SaveChangesAsync();
+                return new ResultModel { Id = person.Id, Success = true };
             }
-            await _context.SaveChangesAsync();
-            _context.Persons.Remove(person);
-            await _context.SaveChangesAsync();
-            return new ResultModel { Id = person.Id, Success = true };
+            return new ResultModel { Success = false };
+
+
         }
 
         public async Task<GetPersonDTO> GetPerson(ItemDto request)
@@ -76,11 +83,64 @@ namespace Assessment.Phonebook.Services.Services
             }
             return null;
         }
+        #endregion
+
+        #region PersonDetail
+        public async Task<ResultModel> CreatePersonDetail(PersonDetailDTO request)
+        {
+            var personDetail = new PersonDetail
+            {
+                PhoneNumber = request.PhoneNumber,
+                MailAddress = request.MailAddress,
+                Location = request.Location,
+                PersonId = request.PersonId
+            };
+
+            await _context.PersonDetails.AddAsync(personDetail);
+            await _context.SaveChangesAsync();
+
+            return new ResultModel { Id = personDetail.Id, Success = true };
+        }
+
+        public async Task<ResultModel> DeletePersonDetail(ItemDto request)
+        {
+
+            var personDetail = _context.PersonDetails.Find(request.Id);
+            if (personDetail != null)
+            {
+                _context.PersonDetails.Remove(personDetail);
+                await _context.SaveChangesAsync();
+                return new ResultModel {Success = true };
+            }
+            return new ResultModel { Success = false };
+        }
+
+        public async Task<GetPersonDTO> GetPersonDetail(ItemDto request)
+        {
+            var person = await _context.Persons.Include(person => person.PersonDetail).FirstOrDefaultAsync(person => person.Id == request.Id);
+            if (person != null)
+            {
+                var personDto = _mapper.Map<GetPersonDTO>(person);
+                return personDto;
+            }
+            return null;
+        }
+        #endregion
     }
     public interface IPersonServices
     {
+        #region Person
         Task<ResultModel> CreatePerson(PersonDto request);
         Task<ResultModel> DeletePerson(ItemDto request);
         Task<GetPersonDTO> GetPerson(ItemDto request);
+        #endregion
+
+        #region PersonDetail
+        Task<ResultModel> CreatePersonDetail(PersonDetailDTO request);
+        Task<ResultModel> DeletePersonDetail(ItemDto request);
+        Task<GetPersonDTO> GetPersonDetail(ItemDto request); 
+        #endregion
+
+
     }
 }
